backend/requirements.txt
# Core ML and Data Processing
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
scipy>=1.8.0
scikit-learn>=1.1.0
pandas>=1.4.0

# EEG Processing
mne>=1.0.0
eeglabio>=0.0.2
PyWavelets>=1.3.0

# Backend Framework
flask>=2.2.0
flask-socketio>=5.3.0
flask-cors>=3.0.0
python-socketio>=5.7.0

# Data Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Signal Processing
librosa>=0.9.0
spectrum>=0.8.0

# Utilities
tqdm>=4.64.0
joblib>=1.1.0
h5py>=3.6.0
python-dotenv>=0.19.0

# Development and Testing
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
